+++ src/app/shared/models/favorites.ts
@@ -4,3 +4,13 @@ export interface Favorite extends Movie {
   description: string;
   rating: number;
 }
+
+export interface FavoriteParams {
+  favorite: Favorite;
+  action: ActionType;
+}
+
+export enum ActionType {
+  UPDATE,
+  DELETE,
+}