+++ src/app/snackbar-annotated/snackbar-annotated.component.ts
@@ -1,0 +1,31 @@
+import { Component, Inject, inject } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {
+  MAT_SNACK_BAR_DATA,
+  MatSnackBarAction,
+  MatSnackBarActions,
+  MatSnackBarLabel,
+  MatSnackBarRef,
+} from '@angular/material/snack-bar';
+import { MatButtonModule } from '@angular/material/button';
+import { TranslateModule } from '@ngx-translate/core';
+
+@Component({
+  selector: 'app-snackbar-annotated',
+  standalone: true,
+  imports: [
+    CommonModule,
+    MatButtonModule,
+    MatSnackBarLabel,
+    MatSnackBarActions,
+    MatSnackBarAction,
+    TranslateModule,
+  ],
+  templateUrl: './snackbar-annotated.component.html',
+  styleUrl: './snackbar-annotated.component.scss',
+})
+export class SnackbarAnnotatedComponent {
+  snackBarRef = inject(MatSnackBarRef);
+
+  constructor(@Inject(MAT_SNACK_BAR_DATA) public data: any) {}
+}